moving iterateTiles() to TileEngine.h
...to use in external files
