+++ src/Battlescape/TileEngine.cpp
@@ -198,36 +198,6 @@ bool calculateParabolaHelper(const Position& origin, const Position& target, dou
 	return false;
 }
 
-/**
- * Iterate through some subset of map tiles.
- * @param save Map data.
- * @param gs Square subset of map area.
- * @param func Call back.
- */
-template<typename TileFunc>
-void iterateTiles(SavedBattleGame* save, MapSubset gs, TileFunc func)
-{
-	const auto totalSizeX = save->getMapSizeX();
-	const auto totalSizeY = save->getMapSizeY();
-	const auto totalSizeZ = save->getMapSizeZ();
-
-	gs = MapSubset::intersection(gs, MapSubset{ totalSizeX, totalSizeY });
-	if (gs)
-	{
-		for (int z = 0; z < totalSizeZ; ++z)
-		{
-			auto rowStart = save->getTile(Position{ gs.beg_x, gs.beg_y, z });
-			for (auto stepsY = gs.size_y(); stepsY != 0; --stepsY, rowStart += totalSizeX)
-			{
-				auto curr = rowStart;
-				for (auto stepX = gs.size_x(); stepX != 0; --stepX, curr += 1)
-				{
-					func(curr);
-				}
-			}
-		}
-	}
-}
 
 /**
  * Generate square subset of map using position and radius.
